casync-http: Dont forget to call PROCESS_UNTIL_WRITTEN
While working on this code, I stumbled on cases where casync got stuck
because we forgot to call PROCESS_UNTIL_WRITTEN here. Well, TBH as long
as we download chunks, we're fine because we end up calling
PROCESS_UNTIL_WRITTEN afterwards.

But in any case, it seems more correct to sync after downloading these
files, and it doesn't hurt.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>