shell-completion: autocomplete subtree path for list commands
Use function __casync_get_arg_n to get the archive or archive index from
the already entered command line, and autocomplete the second argument
for the list commands: list, mtree and stat.

It introduces the helper function __casync_list that performs file and
directory completion from an archive or an archive index.