calocation: fix ESTALE when extracting to regular file
There is an issue when an archive is extracted to a regular file that
exists already.

	# Extract the archive.caibx to dest.bin (dest.bin does not exist yet)
	$ casync extract http://localhost/archive.caibx dest.bin

	# Extract the archive.caibx to existing dest.bin
	$ casync extract http://localhost/archive.caibx dest.bin --log-level debug --verbose
	Acquiring http://localhost/archive.caibx...
	Setting min/avg/max chunk size: 16384 / 65536 / 262144
	Failed to run decoder step: Stale file handle
	Failed to run synchronizer: Stale file handle

The error raised is a difference in the modify time.

In the case of the destination regular file exists (ie. the second call
in the example above), the path goes to several fallocate/write calls
that causes the modification time to be updated and leads to that error.

This commit drops the check for the modification time.