casync-http: remove rr parameter from acquire_file
This commit is the third commit of a patchset of 4 that intends to
factorize the code dedicated to download file using the easy interface
of libcurl.

This commit removes the parameter rr from the function acquire_file() to
make that function re-usable for acquiring chunk files. That function is
now dedicated to acquire a file using the cURL easy interface.