casync-http: set curl verbosity
Enable cURL verbosity for the notice and higher log levels (ie. debug).

When CURLOPT_VERBOSE is enabled, libcurl outputs traces to stderr, such
as headers and a lot of internal information.

Now, the libcurl verbose informations are dumped to stdout when the
options --log-level notice or debug are specified to casync.